requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
dj-database-url==2.3.0
Django==5.1.3
django-allauth==65.2.0
fonttools==4.55.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
oauthlib==3.2.2
packaging==24.2
paho-mqtt==2.1.0
pillow==11.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
whitenoise==6.8.2
